throw error when study code does not exist
