+++ constants/error-constants.js
@@ -1,0 +1,11 @@
+exports.errorName = {
+  STUDY_CODE_NOT_FOUND: "STUDY_CODE_NOT_FOUND",
+};
+
+exports.errorType = {
+  STUDY_CODE_NOT_FOUND: {
+    message:
+      "The provided study code does not match any existing ICDC studies.",
+    statusCode: 400,
+  },
+};
