+++ data-management/init-graphql.js
@@ -1,6 +1,7 @@
 const { buildSchema } = require("graphql");
 const { graphqlHTTP } = require("express-graphql");
 const { mapCollectionsToStudies } = require("./data-interface");
+const { formatErrorResponse } = require("../util/error-util");
 
 const schema = buildSchema(
   require("fs").readFileSync("graphql/schema.graphql", "utf8")
@@ -18,22 +19,7 @@ module.exports = graphqlHTTP((req, res) => {
     rootValue: root,
     context: { req },
     customFormatErrorFn: (error) => {
-      // TODO: improve/enhance app error handling
-      // code in try block below won't throw error
-      // need to throw specific errors in data-interface
-      let status;
-      let body = { error: undefined };
-      try {
-        status = 400;
-        body.error =
-          error.message.replace(/\"/g, "") +
-          ` ${JSON.stringify(error.locations).replace(/\"/g, "")}`;
-      } catch (err) {
-        status = 500;
-        body.error = "Internal server error: " + error.message;
-      }
-      res.status(status);
-      return body;
+      return formatErrorResponse(res, error);
     },
   };
 });
